Updated .gitignore and added deployment-script for ethereum holesky
For deployment on the Ethereum Holesky-Testnet first the live data of the sync-commitee and the slots has to be requested. Therefore the free API of https://beaconcha.in/api/v1/docs/index.html is used. Therefore the deployment-script requests the data and deploys the relay-512-contract with it. To deploy the relay at the testnet  the specific routes and parameters are set in the truffle-config.