adding new validation-functions of "old" ETH-Relay-Implementation and further block-validation
Based on the implementation of transaction validation in https://github.com/pantos-io/research-ethrelay methods are transfered. Furthermore the "validate"-method is changed in the way that it is added by a block-validation based on a call of bls-signature-validation based on "alt_bn128". At the moment it is mocked due to missing precompile on holesky-testnet. Further parameters are added to transaction-validation call including: bytes32 blockhash, bool[SYNC_COMMITTEE_SIZE] memory participants.